wallet: move `ImportDescriptors`/`FuzzedWallet` to util
